+++ Source/ACE.Database/ISerializedShardDatabase.cs
@@ -30,7 +30,7 @@ namespace ACE.Database
 
         void DeleteOrRestore(ulong unixTime, uint id, Action<bool> callback);
 
-        void GetMaxPlayerId(Action<uint> callback);
+        void GetCurrentId(uint min, uint max, Action<uint> callback);
 
         void SetCharacterAccessLevelByName(string name, AccessLevel accessLevel, Action<uint> callback);